Switch field used to differentiate episodes
The source plugin used to pull podcast episodes from a feed uses the
"link" value by default. This was working when episodes were edited to
include this field, but broke when new episodes were uploaded with empty
values for that field.

This change uses the "guid" value, which is automatically generated by
the hosting provider and should be guaranteed to be present.