Only render Spotify player when Spotify URL used
Logic previousy assumed that a Spotify link would always be found in the
description, breaking if it was not included.

This moves the variables and component inside conditionals based on
whether any matches were found via the regular expression.

The variables were switched from const to var becase otherwise they are
not available outside of the conditional block.